Refactored app.rb to handle books fuctions
